Boolti-235 feat: QR FullScreen 을 TicketDetailScreen 로 이동 N매의 티켓을 티켓 상세 화면과 QR 화면이 공유하기 위함
