+++ presentation/src/main/java/com/nexters/boolti/presentation/screen/qr/QrFullScreen.kt
@@ -12,6 +12,7 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -22,18 +23,44 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.NavBackStackEntry
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
 import com.nexters.boolti.presentation.R
+import com.nexters.boolti.presentation.screen.MainDestination
+import com.nexters.boolti.presentation.screen.data
+import com.nexters.boolti.presentation.screen.ticket.detail.TicketDetailViewModel
+import com.nexters.boolti.presentation.screen.ticketName
 import com.nexters.boolti.presentation.theme.Grey10
 import com.nexters.boolti.presentation.theme.Grey85
 import com.nexters.boolti.presentation.theme.Grey90
 import com.nexters.boolti.presentation.util.rememberQrBitmapPainter
 
+fun NavGraphBuilder.QrFullScreen(
+    popBackStack: () -> Unit,
+    getSharedViewModel: @Composable (NavBackStackEntry) -> TicketDetailViewModel,
+    modifier: Modifier = Modifier,
+) {
+    composable(
+        route = "${MainDestination.Qr.route}/{$data}?ticketName={$ticketName}",
+        arguments = MainDestination.Qr.arguments,
+    ) { entry ->
+        QrFullScreen(
+            modifier = modifier,
+            viewModel = getSharedViewModel(entry),
+        ) { popBackStack() }
+    }
+}
+
 @Composable
 fun QrFullScreen(
     modifier: Modifier = Modifier,
-    viewModel: QrFullViewModel = hiltViewModel(),
+    viewModel: TicketDetailViewModel = hiltViewModel(),
     onClose: () -> Unit,
 ) {
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+
     ConstraintLayout(
         modifier = modifier
             .background(Color.White)
@@ -70,7 +97,7 @@ fun QrFullScreen(
         ) {
             Text(
                 modifier = Modifier.padding(top = 16.dp),
-                text = viewModel.ticketName,
+                text = uiState.ticketGroup.ticketName,
                 style = MaterialTheme.typography.titleMedium,
                 color = Grey85.copy(alpha = .85f),
             )
@@ -81,7 +108,7 @@ fun QrFullScreen(
                     .background(Color.White)
                     .padding(14.dp),
                 painter = rememberQrBitmapPainter(
-                    viewModel.data,
+                    content = uiState.ticketGroup.tickets.first().entryCode,
                     size = 260.dp,
                 ),
                 contentScale = ContentScale.Inside,
